% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cldr_examples.R
\docType{data}
\name{cldr_examples}
\alias{cldr_examples}
\title{List of CLDR country name codes and associated examples}
\format{
data frame
}
\usage{
cldr_examples
}
\description{
\itemize{
\item Code: CLDR code
\item Example: French Southern Territories in different languages
}
}
\keyword{datasets}
