% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_fields.R
\name{closest_field}
\alias{closest_field}
\title{Accepts a vector of unique values and returns an ordered data.frame or tibble
that shows the most closely matching fields in the codelist data.frame}
\usage{
closest_field(x)
}
\arguments{
\item{x}{A vector of unique codes or names to check against the codelist data.frame}
}
\description{
Accepts a vector of unique values and returns an ordered data.frame or tibble
that shows the most closely matching fields in the codelist data.frame
}
\examples{
# Most likely 2 character code fields
closest_field(c('CM','AL','PA','MK','HA','CO','ZI','KU','TH','FI'))
# Most likely country name fields
closest_field(c('Ivory Coast','Guinea','Iran','Russia','North Korea'))
}
