% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist_panel.R
\docType{data}
\name{codelist_panel}
\alias{codelist_panel}
\title{Country Code Translation Data Frame (Country-Year Panel)}
\format{
data frame with codes as columns
}
\usage{
codelist_panel
}
\description{
A panel of country-year observations with various codes
}
\keyword{datasets}
