\name{countrycode-package}
\alias{countrycode-package}
\docType{package}
\title{
Countrycode is a utility to convert country coding schemes
}
\description{The countrycode() function can convert to and from 7 different country coding schemes. It uses regular expressions to convert long country names (e.g. Sri Lanka) into any of those coding schemes, or into standardized country names (official short English). It can create variables with the name of the continent and/or region to which each country belongs. The function countrycode.nomatch() returns a list of conversions that cannot be completed using the package countrycode. The function countryframe() can generate data frames composed of all directed or undirected dyad-year, or country-year observations.
}
\details{
\tabular{ll}{
Package: \tab countrycode\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2010-08-14\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\author{
Vincent Arel-Bundock <varel@umich.edu>
}

\keyword{ package }

