% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrycode_factory.R
\name{countrycode_factory}
\alias{countrycode_factory}
\title{create new `countrycode` functions with alternative dictionaries and
defaults}
\usage{
countrycode_factory(
  origin = "iso3c",
  destination = "country.name",
  warn = FALSE,
  nomatch = NA,
  custom_dict = NULL,
  custom_match = NULL,
  origin_regex = NULL
)
}
\description{
create new `countrycode` functions with alternative dictionaries and
defaults
}
