% Generated by roxygen2 (4.0.1): do not edit by hand
\name{countrycode_test}
\alias{countrycode_test}
\title{countrycode tests}
\usage{
countrycode_test()
}
\description{
Runs a series of tests.
Returns TRUE if all tests passed.
Prints test description when some tests fail.
}
\keyword{countrycode_test}

