% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryname_dict.R
\docType{data}
\name{countryname_dict}
\alias{countryname_dict}
\title{A dataframe of alternative country names in many languages. Used internally by
the \code{countryname} function.}
\format{
dataframe
}
\description{
A dataframe of alternative country names in many languages. Used internally by
the \code{countryname} function.
}
\keyword{datasets}
