% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrycode-package.R
\docType{package}
\name{countrycode-package}
\alias{countrycode-package}
\title{Convert Country Codes or Country Names}
\description{
Convert country codes or country names
}
\details{
The \code{countrycode} function can convert to and from several different
country coding schemes.  It uses regular expressions to convert country
names (e.g. Sri Lanka) into any of those coding schemes, or into
standardized country names in several languages.  It can create variables
with the name of the continent and/or several regional groupings to which
each country belongs.

Type ?codelist to get a list of available origin and destination
codes.
}
\references{
\url{http://arelbundock.com}
    \url{https://github.com/vincentarelbundock/countrycode}
}
\author{
Vincent Arel-Bundock \email{vincent.arel-bundock@umontreal.ca}
}
\keyword{package}
