% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrycode-package.R
\docType{package}
\name{countrycode-package}
\alias{countrycode-package}
\title{Convert Country Codes or Country Names}
\description{
Convert country codes or country names
}
\details{
The \code{countrycode} function can convert to and from 12 different country coding schemes.
It uses regular expressions to convert long country names (e.g. Sri Lanka) into any of
those coding schemes, or into standardized country names (official short English).
It can create variables with the name of the continent and/or region to which each
country belongs.
}
\author{
Vincent Arel-Bundock \email{varel@umich.edu}
}
\references{
\url{http://www.umich.edu/~varel}
    \url{https://github.com/vincentarelbundock/countrycode}
}
\keyword{package}

