\docType{package}
\name{countrycode-package}
\alias{countrycode-package}
\title{Convert country codes or country names}
\description{
  Convert country codes or country names
}
\details{
  The \code{countrycode} function can convert to and from 7
  different country coding schemes. It uses regular
  expressions to convert long country names (e.g. Sri
  Lanka) into any of those coding schemes, or into
  standardized country names (official short English). It
  can create variables with the name of the continent
  and/or region to which each country belongs.
}
\author{
  Vincent Arel-Bundock \email{varel@umich.edu}
}
\references{
  \url{http://www.umich.edu/~varel}
  \url{https://github.com/vincentarelbundock/countrycode}
}
\keyword{package}

