% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrycode.R
\name{countrycode}
\alias{countrycode}
\title{Convert Country Codes}
\usage{
countrycode(sourcevar, origin, destination, warn = FALSE, dictionary = NULL,
  extra = NULL)
}
\arguments{
\item{sourcevar}{Vector which contains the codes or country names to be converted}

\item{origin}{Coding scheme of origin (name enclosed in quotes "")}

\item{destination}{Coding scheme of destination (name enclosed in quotes "")}

\item{warn}{Prints unique elements from sourcevar for which no match was found}

\item{dictionary}{A data frame which supplies custom country codes.
Variables correspond to country codes, observations must refer to unique
countries.  When countrycode uses a user-supplied dictionary, no sanity
checks are conducted. The data frame format must resemble
countrycode::countrycode_data. Custom dictionaries only work with strings
(no regexes).}

\item{extra}{A data frame which supplies additional country codes in the scheme
chosen by origin/destination, to supplement the official list. Must be a two-
column data frame or a list of two vectors. Column names must match if used. 
Warnings will be suppressed if a match is returned from this data frame.
Regexes not supported.}
}
\description{
Converts long country names into one of many different coding schemes.
Translates from one scheme to another. Converts country name or coding
scheme to the official short English country name. Creates a new variable
with the name of the continent or region to which each country belongs.
}
\note{
Supports the following coding schemes: Correlates of War character,
  CoW-numeric, ISO3-character, ISO3-numeric, ISO2-character, IMF numeric, International
  Olympic Committee, FIPS 10-4, FAO numeric, United Nations numeric,
  World Bank character, official English short country names (ISO), continent, region.

  The following strings can be used as arguments for \code{origin} or
  \code{destination}: "cowc", "cown", "iso3c", "iso3n", "iso2c", "imf",
  "fips104", "fao", "ioc", "un", "wb", "country.name".  The following strings can be
  used as arguments for \code{destination} \emph{only}:  "continent", "region",
  "eu28", "ar5"
}
\examples{
codes.of.origin <- countrycode::countrycode_data$cowc # Vector of values to be converted
countrycode(codes.of.origin, "cowc", "iso3c")

two_letter <- c("AU", "US", "XK") # World Bank uses user-assigned XK for Kosovo
countrycode(two_letter, "iso2c", "country.name", warn=TRUE)
countrycode(two_letter, "iso2c", "country.name", warn=TRUE,
            extra=list(c("XK", "JG"),c("Kosovo", "Channel Islands")))
}
\keyword{countrycode}

